introduce decision actions
A decision action is an action that does not declare a job
(or its job is null) in the workflow action definition.
Hence these do not start a Nomad job.
All their "work" is done during workflow evaluation.

When a decision action is run its success facts are published
immediately because a finished evaluation is what defines success
for a decision action. Decision actions cannot fail at runtime
because there is no runtime so they have no failure facts
(only an error at evaluation time could be considered failure).