Clean up the Nix evaluator
This now first transforms the environment variables to an attrset using
`nix-instantiate`, and then passes it to `nix eval`. This allows for use
of the evaluation cache and IMHO cleaner code.
The script is now also in its own file so we don't have to worry about
two stages of interpolation and get instant shellcheck feedback.
Also fixed a few shellcheck issues and use the awesome
`writeShellApplication`.