cache input matching
Save action active status in separate table.
This makes it impossible to reactivate previous versions of an action.
This is done this way because
previous versions are not matched with on fact creation
so on reactivation they would have to be matched with all facts
from the DB that were created since the old version was deactivated.
Alternatively we could always match with all versions of all actions
but that would be a waste of resources
as most old versions are never reactivated.

Also add `/action/current/{name}` endpoint.